requirements.txt
openai
pyttsx3
SpeechRecognition
pyaudio
requests
python-dotenv
beautifulsoup4
lxml
schedule

